refactor: update props names
